+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Released: upcoming
 
  * ```[ADReCS]``` #60 Add ADReCS data source module
  * ```[UNII]``` #29 Fix UNII updater for new website
+ * ```[HGNC]``` #13 Remove previous symbols and add ensembl gene id from HGNC gene mapping
 
 ### 🔧 Bug fixes and other changes